Send 0L usage after successfully publishing the actual usage
Since messages are retained the last message in the state topic will
contain the actual usage. This when combined with Home Assistant
restarting the next day before the actual usage of the coming day
published to set the usage for the day incorrectly to the same value as
the day before.
Publishing 0 after the actual usage won't change the calculation as for
that a 0 value is already needed.