Implement waking up just before midnight
Use NTP to fetch current time before going to deep sleep and use that to
calculate the amount of time to sleep in order to wake up at the right
time.

If querying the NTP server fails fall back to just sleeping for a day.

By setting the timezone with POSIX format we can avoid any network
traffic other than actual NTP queries on 123/UDP.